requirements.txt
alembic==1.16.1
azure-core==1.34.0
azure-storage-blob==12.25.1
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.3
Flask==3.1.1
Flask-Bcrypt==1.0.1
flask-cors==6.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
google-cloud-storage==2.14.0 # New
# google-cloud-aiplatform==1.49.0 # Optional, uncomment if you implement real Vertex AI SDK calls
gunicorn==23.0.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.41
tomli==2.2.1
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==3.1.3
firebase-admin==6.1.0
